pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "policytester"
version = "0.1.0"
description = "Test kubernetes network policies"
readme = "README.md"
authors = [{ name = "Erik Brakkee", email = "erik@brakkee.org" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Testing",
    "Topic :: System :: Networking",
    "Topic :: System :: Networking :: Firewalls",
]
keywords = ["kubernetes", "network", "policy", "network policy"]
dependencies = [
    "kubernetes >= 23.6.0",
    "cerberus >= 1.3.4",
    "pyyaml >= 6.0",
    "attrdict >= 2.0.1",
    'tomli; python_version < "3.11"',
]
requires-python = ">=3.8"

[project.optional-dependencies]
dev = ["black", "bumpver", "isort", "pip-tools", "pytest"]

[project.urls]
Homepage = "https://github.com/ErikEngerd/policytester"

[project.scripts]
policytester = "policytester.__main__:main"


[tool.bumpver]
current_version = "0.1.0"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message = "bump version {old_version} -> {new_version}"
commit = true
tag = true
push = true

[tool.bumpver.file_patterns]
"pyproject.toml" = [
    'current_version = "{version}"',
]
"src/policytester/__init__.py" = [
    '__version__ = "{version}"'
]


